Move shutdown signal handler to judoscale core init
While this is related to the reporter and the fact it is a thread we
need to gracefully shutdown, it's not specific to the reporter and more
to the core package itself.

Plus, we'll probably need to extend this to shutdown the utilization
tracker thread soon, so it makes sense to put it in a more "global"
place.
